# src/user_cred/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "DSTI DevOps Project",
        "version": "1.0.0",
        "title": "Crud Flask API Tutorial",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "tags": [
        {
            "name": "Products API",
            "description": "Test API"
        }
    ],
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "schemas": {
            "users": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "Id": {
                "type": "integer",
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                }
            },
            "update": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
        "/users": {
            "post": {
                "tags": [
                    "Add Users to API"
                ],
                "summary": "Add user",
                "requestBody": {
                    "description": "Post New user To Database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/users"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Get All Users from API"
                ],
                "summary": "Get All users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Get user from API"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to get",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Update user in API"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/update"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete user from API"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to delete",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}